refactor: dto 클래스 분리
